refactor: PartitionsSourceConfig is part of local scheduler
This is specifically about a configuration only used during the scheduling process. It is later used in the seeding of the initial VEc<PartitionIds> provided into the PartitionsSource trait implementations.
