+++ compactor/src/components/partitions_source/scheduled.rs
@@ -1,18 +1,17 @@
 use std::{fmt::Display, sync::Arc};
 
 use async_trait::async_trait;
-use compactor_scheduler::{temp::PartitionsSourceConfig, CompactionJob, Scheduler};
+use compactor_scheduler::{CompactionJob, Scheduler};
 use data_types::{PartitionId, PartitionsSource};
 
 #[derive(Debug)]
 pub struct ScheduledPartitionsSource {
     scheduler: Arc<dyn Scheduler>,
-    config: PartitionsSourceConfig,
 }
 
 impl ScheduledPartitionsSource {
-    pub fn new(scheduler: Arc<dyn Scheduler>, config: PartitionsSourceConfig) -> Self {
-        Self { scheduler, config }
+    pub fn new(scheduler: Arc<dyn Scheduler>) -> Self {
+        Self { scheduler }
     }
 }
 
@@ -25,7 +24,7 @@ impl Display for ScheduledPartitionsSource {
 #[async_trait]
 impl PartitionsSource for ScheduledPartitionsSource {
     async fn fetch(&self) -> Vec<PartitionId> {
-        let job: Vec<CompactionJob> = self.scheduler.get_job(&self.config).await;
+        let job: Vec<CompactionJob> = self.scheduler.get_job().await;
         job.into_iter().map(|job| job.partition_id).collect()
     }
 }
