+++ compactor/src/components/report.rs
@@ -26,7 +26,6 @@ pub fn log_config(config: &Config) {
         percentage_max_file_size,
         split_percentage,
         partition_timeout,
-        partitions_source,
         shadow_mode,
         ignore_partition_skip_marker,
         min_num_l1_files_to_compact,
@@ -61,7 +60,6 @@ pub fn log_config(config: &Config) {
         percentage_max_file_size,
         split_percentage,
         partition_timeout_secs=partition_timeout.as_secs_f32(),
-        %partitions_source,
         shadow_mode,
         ignore_partition_skip_marker,
         min_num_l1_files_to_compact,
