+++ compactor_scheduler/src/lib.rs
@@ -19,7 +19,7 @@
 #![allow(clippy::missing_docs_in_private_items)]
 
 mod local_scheduler;
-pub use local_scheduler::LocalScheduler;
+pub use local_scheduler::{LocalScheduler, PartitionsSourceConfig};
 mod remote_scheduler;
 mod scheduler;
 pub use scheduler::*;
@@ -40,7 +40,10 @@ pub async fn create_compactor_scheduler_service(
     match scheduler_config.compactor_scheduler_type {
         CompactorSchedulerType::Local => {
             let shard_config = ShardConfig::from_config(scheduler_config.shard_config);
+            let partitions_source_config =
+                PartitionsSourceConfig::from_config(scheduler_config.partition_source_config);
             Arc::new(LocalScheduler::new(
+                partitions_source_config,
                 catalog,
                 BackoffConfig::default(),
                 None,
@@ -56,8 +59,3 @@ pub async fn create_compactor_scheduler_service(
         }
     }
 }
-
-// temporary mod, for this commit only.
-// code still being consumed in the compactor, by direct function call.
-// TODO: move into LocalScheduler, which is used by the grpc service.
-pub mod temp;
