+++ compactor_scheduler/src/remote_scheduler.rs
@@ -5,10 +5,7 @@ use async_trait::async_trait;
 use generated_types::influxdata::iox::compactor_scheduler::v1::*;
 use tonic::{transport::Channel, Response, Status};
 
-use crate::{
-    scheduler::{CompactionJob, Scheduler},
-    temp::PartitionsSourceConfig,
-};
+use crate::scheduler::{CompactionJob, Scheduler};
 
 /// Implementation of the [`Scheduler`] for remote coordinated scheduling.
 #[derive(Debug)]
@@ -31,7 +28,7 @@ impl RemoteScheduler {
 
 #[async_trait]
 impl Scheduler for RemoteScheduler {
-    async fn get_job(&self, _: &PartitionsSourceConfig) -> Vec<CompactionJob> {
+    async fn get_job(&self) -> Vec<CompactionJob> {
         let mut client = self.client.clone();
         let response: Result<Response<GetJobResponse>, Status> =
             client.get_job(GetJobRequest {}).await;
