+++ compactor_scheduler/src/scheduler.rs
@@ -4,8 +4,6 @@ use async_trait::async_trait;
 use data_types::PartitionId;
 use generated_types::influxdata::iox::compactor_scheduler::v1::{self as proto};
 
-use crate::temp::PartitionsSourceConfig;
-
 /// Job assignment for a given partition.
 #[derive(Debug)]
 pub struct CompactionJob {
@@ -32,5 +30,5 @@ impl From<proto::CompactionJob> for CompactionJob {
 #[async_trait]
 pub trait Scheduler: Send + Sync + Debug + Display {
     /// Get partitions to be compacted.
-    async fn get_job(&self, _: &PartitionsSourceConfig) -> Vec<CompactionJob>;
+    async fn get_job(&self) -> Vec<CompactionJob>;
 }
